自定义初始化

+++ install.sh
@@ -1,18 +1,9 @@
 #!/bin/sh
 source basic.sh
 
-sudo ./install-steps/macos.sh
+# sudo ./install-steps/macos.sh
 
 brew install python3
-pip3 install shadowsocks
-
-# install and use shadowsocks
-if not_tt_network; then
-    nohup sslocal -q -c ~/.macbootstrap/tools/netconf &> /private/tmp/nohup.out&
-    export ALL_PROXY=socks5://127.0.0.1:14179
-else
-    echo "You are in toutiao network, no need to use ss now"
-fi
 
 if [[ ! -e /Applications/iTerm.app ]]; then
     brew cask install iterm2
@@ -26,11 +17,11 @@ else
     echo "You have installed iTerm2"
 fi
 
-if [[ ! -e /Applications/SourceTree.app ]]; then
-    brew cask install sourcetree
-else
-    echo "You have installed SourceTree"
-fi
+# if [[ ! -e /Applications/SourceTree.app ]]; then
+#     brew cask install sourcetree
+# else
+#     echo "You have installed SourceTree"
+# fi
 
 if [[ ! -e /Applications/WeChat.app ]]; then
     brew cask install wechat
@@ -78,29 +69,29 @@ else
     echo "You have installed coreutils"
 fi
 
-brew install --HEAD universal-ctags/universal-ctags/universal-ctags
-brew install redis
-brew_install cmake
-brew_install gawk
-brew_install autojump
-brew_install wget
-brew_install nvm
-brew_install exiv2
-brew_install ssh-copy-id
-brew_install imagemagick
-brew_install catimg
-brew_install gpg
-brew_install icdiff
-brew_install scmpuff
-brew_install fzf
-brew_install fd
-brew_install the_silver_searcher
-brew_install nvim
-brew_install exiftool
-brew_install archey
-brew_install ranger
-brew_install git-lfs && git lfs install
-$(brew --prefix)/opt/fzf/install --all
+# brew install --HEAD universal-ctags/universal-ctags/universal-ctags
+# brew install redis
+brew install cmake
+# brew install gawk
+brew install autojump
+brew install wget
+brew install nvm
+brew install exiv2 #用来提取图片元信息
+brew install ssh-copy-id
+brew install imagemagick
+# brew install catimg
+# brew install gpg
+brew install icdiff
+# brew install scmpuff
+# brew install fzf
+brew install fd
+# brew install the_silver_searcher
+brew install neovim
+# brew install exiftool
+brew install archey
+brew install ranger
+# brew install git-lfs && git lfs install
+# $(brew --prefix)/opt/fzf/install --all
 
 # link git config
 mv ~/.gitconfig ~/.gitconfig_backup
@@ -116,17 +107,21 @@ fi
 backup_file ~/.zshrc
 git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
 git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+
 ln -s ~/.macbootstrap/zsh-config/.zshrc ~/.zshrc
 
 # vim configuration
-backup_file ~/.vim
-backup_file ~/.config/nvim/
-git clone https://github.com/bestswifter/vim-config.git ~/.config/nvim
-ln -s ~/.config/nvim ~/.vim
+if [[ ! -e  ~/.config/nvim]]; then
+    backup_file ~/.vim
+    backup_file ~/.config/nvim/
+    git clone https://github.com/rafi/vim-config.git ~/.config/nvim
+    ln -s ~/.config/nvim ~/.vim
+fi
 
 # ESLint configuration
 backup_file ~/.eslintrc.js
 backup_file ~/.eslintrc
+
 ln -s ~/.macbootstrap/.eslintrc.js ~/.eslintrc.js
 
 # Ranger configuration
@@ -140,11 +135,9 @@ backup_file "$old_rc_conf"
 ln -s ~/.macbootstrap/config/ranger/commands.py "$old_commands_py"
 ln -s ~/.macbootstrap/config/ranger/rc.conf "$old_rc_conf"
 
-./install-steps/dependencies.before.sh
-
-unset ALL_PROXY
-./install-steps/dependencies.after.sh
-./install-steps/sogou_sync.sh
+# ./install-steps/dependencies.before.sh
+# ./install-steps/dependencies.after.sh
+# ./install-steps/sogou_sync.sh
 
 # ssh configuration
 backup_file ~/.ssh/config
@@ -154,6 +147,6 @@ fi
 ln -s ~/.macbootstrap/zsh-config/ssh_config ~/.ssh/config
 
 # Personal
-./install-steps/personal.sh
+# ./install-steps/personal.sh
 ./personal.sh
